requirements.txt
djangorestframework
amqp==2.4.2
backcall==0.1.0
billiard==3.6.0.0
celery==4.3.0
certifi==2019.3.9
chardet==3.0.4
decorator==4.4.0
Django==2.2.1
django-cors-headers==3.0.1
django-debug-toolbar==1.11
django-storages==1.7.1
idna==2.8
ipython==7.5.0
ipython-genutils==0.2.0
jedi==0.13.3
kombu==4.5.0
numpy==1.16.3
pandas==0.24.2
parso==0.4.0
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.0.0
prompt-toolkit==2.0.9
ptyprocess==0.6.0
Pygments==2.4.0
python-dateutil==2.8.0
python-decouple==3.1
python-http-client==3.1.0
pytz==2019.1
qrcode==6.1
raven==6.10.0
requests==2.21.0
sendgrid==6.0.5
six==1.12.0
sqlparse==0.3.0
traitlets==4.3.2
urllib3==1.24.3
vine==1.3.0
wcwidth==0.1.7
